refactor: split product details page in multiple components
- Created multiple components to better optimize the product
details page.

- Removed unecessary use client statements from the main page

- Removed unecessary use state and use effect hooks

- Created client components where needed
