+++ web/app/(home)/offer-slider.tsx
