+++ web/package.json
@@ -17,7 +17,8 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-slick": "0.30.3",
-    "slick-carousel": "1.8.1"
+    "slick-carousel": "1.8.1",
+    "tailwind-merge": "^3.0.2"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.9.4",
